services/setfit_inference/requirements.txt
fastapi
uvicorn
pydantic
setfit
PyYAML
python-multipart
python-jose
typing-extensions
python-dotenv
huggingface-hub
loguru